refactor(core): rename getBlockAlignedChangedRanges to getRebuildRanges and handle pos-only steps
