+++ packages/core/src/helpers/getChangedRanges.ts
@@ -46,19 +46,16 @@ export function getChangedRanges(transform: Transform): ChangedRange[] {
     // e.g. when setting a mark, node attribute, etc.
     // @ts-ignore
     if (!stepMap.ranges.length) {
-      const { from, pos, to } = steps[index] as Step & {
+      const { from, to } = steps[index] as Step & {
         from?: number
-        pos?: number
         to?: number
       }
 
-      if (from !== undefined && to !== undefined) {
-        ranges.push({ from, to })
-      } else if (pos !== undefined) {
-        ranges.push({ from: pos, to: pos + 1 })
-      } else {
+      if (from === undefined || to === undefined) {
         return
       }
+
+      ranges.push({ from, to })
     } else {
       stepMap.forEach((from, to) => {
         ranges.push({ from, to })
